.history/requirements_20240605160046.txt
django-channels 
daphne 
channels-redis 
celery
redis